Refactor as constructor funcs
